+++ packages/plugin-packages/mashroom-portal-demo-composite-app/src/js/App.tsx
@@ -1,35 +1,16 @@
 
-import React, {useEffect, useMemo, useRef} from 'react';
+import React, {useEffect, useRef} from 'react';
 import PrivateMessageBus from './PrivateMessageBus';
+import DIALOG from './dialog';
+import {getDialogElementId} from './utils';
+
 import type {MutableRefObject} from 'react';
 import type {MashroomPortalMessageBus, MashroomPortalAppService} from '@mashroom/mashroom-portal/type-definitions';
 import type {ActiveApp} from './types';
 
-const DIALOG = [
-    {
-        name: 'Mashroom Portal Demo Vue App',
-        appConfig: {
-            message: 'This simple Vue SPA that acts as the first page',
-            pingButtonLabel: 'Next'
-        }
-    },
-    {
-        name: 'Mashroom Portal Demo Angular App',
-        appConfig: {
-            message: 'This simple Angular SPA that acts as the second page',
-            pingButtonLabel: 'Next'
-        }
-    },
-    {
-        name: 'Mashroom Portal Demo Svelte App',
-        appConfig: {
-            message: 'This simple Svelte SPA that acts as the third and last page',
-            pingButtonLabel: 'Start over'
-        }
-    }
-];
-
 type Props = {
+    appId: string;
+    activeApp?: ActiveApp | undefined | null;
     messageBus: MashroomPortalMessageBus;
     portalAppService: MashroomPortalAppService;
 }
@@ -67,13 +48,18 @@ const onPing = (activeAppRef: MutableRefObject<ActiveApp | undefined>, dialogEle
     loadDialogPage(nextIdx, activeAppRef, dialogElementId, portalAppService);
 };
 
-export default ({messageBus, portalAppService}: Props) => {
-    const dialogElementId = useMemo(() => `_${Math.floor(Math.random() * 100000)}`, []);
+export default ({appId, activeApp, messageBus, portalAppService}: Props) => {
+    const dialogElementId = getDialogElementId(appId);
     const activeAppRef = useRef<ActiveApp | undefined>();
 
     useEffect(() => {
-        // Load first page
-        loadDialogPage(0, activeAppRef, dialogElementId, portalAppService);
+        if (activeApp) {
+            // SSR rendered App
+            activeAppRef.current = activeApp;
+        } else {
+            // Load first page
+            loadDialogPage(0, activeAppRef, dialogElementId, portalAppService);
+        }
         // Install a private message bus
         const privateMessageBus = new PrivateMessageBus(messageBus, activeAppRef,
             () => onPing(activeAppRef, dialogElementId, portalAppService));
@@ -86,10 +72,13 @@ export default ({messageBus, portalAppService}: Props) => {
         <div className='mashroom-demo-composite-app'>
             <div className="info">
                 <span className="info-icon" />
-                This composite App uses an Angular, a Vue and a Svelte App to create a dialog.
+                This <strong>Composite App</strong> uses a Vue, an Angular and a Svelte App to create a dialog.
                 While the App itself is written in React.
+                <br/><br/>
+                This App is also capable of server-side rendering itself and the first <em>embedded</em> App.
             </div>
-            <div id={dialogElementId} className="dialog" />
+            { /* dangerouslySetInnerHTML is necessary here to prevent the server-side content of this div to be removed during hydration! */ }
+            <div id={dialogElementId} className="dialog" dangerouslySetInnerHTML={{ __html: '' }} />
         </div>
     );
 };