+++ packages/plugin-packages/mashroom-portal-demo-composite-app/src/js/types.ts
@@ -1,7 +1,10 @@
 
 
 export type ActiveApp = {
-    dialogIdx: number;
-    appId: string;
+    readonly dialogIdx: number;
+    readonly appId: string;
 }
 
+export type SSRPreloadedState = {
+    readonly activeApp: ActiveApp;
+}