+++ packages/plugin-packages/mashroom-portal-demo-composite-app/src/js/utils.ts
@@ -1,0 +1,2 @@
+
+export const getDialogElementId = (appId: string) => `_${appId}_sub_apps1`;